Update:demo测试已经通过，将demo部分代码整合进FurryFusion功能中，将FurryFusion作为FurryModule的子插件而非平行插件。
