+++ src/plugins/FurryModule/FurryFusion/__init__.py
@@ -1,0 +1,1 @@
+from .FurryFusion import *
